requirements.txt
crewai
langchain_google_genai  ## google gemini model
load_dotenv
crewai_tools